Change to parent products in editable kits to auto select a child product
The old code can add the parent SKU to the cart. This change will auto select a child product so the correct SKU is added to the cart.

Note: There is a issue with the [@fixed_assemble@] tag as well. It is always set to one for parent products. Current work around is to add `[%set [@fixed_fixed_assemble@]%][%if [@min_assemble@] eq [@max_assemble@]%]1[%else%]0[%/if%][%/set%]` and replace `[@fixed_assemble@]` with `[@fixed_fixed_assemble@]` in the two references in the parent qty field.