Fix Ajax Search Dropdown Item Selection
Previously clicking an item in ajax search dropdown would make the system search for the name of the product clicked....not sure why it was setup this way but it means if the product name contains bad characters the customer is taken to a no results found search page. This instead takes the customer directly to the product page preventing the above from occurring and stopping customers from having to click twice.