Fixing wrong multi level pricing for variations
If Multilevel Pricing is added to another price group and you login as the other group, it was showing the price of one of the cheapest multilevel price rather than the price for one product.

For example, it you set up Multilevel Pricing on Price B with Min 2 / Max 2 / Price 137.94 and Min 3 / Max 3 / Price 130.68. When you login as a customer assigned to Group B, the single price is showing as $130.68 rather than the regular Price B.

Changing [@price@] to [@store_price@] fixes the problem.