Purging Logos
This change will cause logos to be purged when hitting save on the theme editor. Customers will be able to make sure all browsers see the new logo without needing a design tweak. Alternatively replace [@config:neto_css_version@] with a format function loading the date (only to the day) so that the logo is "purged" daily